test(logger,metrics): move e2eUtils into the same folder as cdk-cli and don't expose it outside of the package
