+++ packages/commons/src/index.ts
@@ -1,5 +1,4 @@
 export * from './utils/lambda';
 export * from './Utility';
-export * from './tests/e2e';
 export * as ContextExamples from './tests/resources/contexts';
 export * as Events from './tests/resources/events';
