--- packages/commons/src/tests/e2e/index.ts
@@ -1,2 +1,0 @@
-export * from './e2eUtils';
-export * from './InvocationLogs';
